+++ app/models/api_operations/notes.rb
@@ -2,7 +2,7 @@ module ApiOperations
 
   module Notes
 
-    def self.synchronize_account(account, new_user_maps)
+    def self.synchronize_account(account, new_user_maps, account_not_synchronized_yet)
       ApiOperations::Common.log(:debug,nil,"Started the synchronization of the notes of the account with id = " + account.id.to_s)
 
       # run a synchronization just for each new user map
@@ -11,7 +11,7 @@ module ApiOperations
       end
       
       # run a normal complete synchronization
-      self.synchronize_account_process(account,nil)
+      self.synchronize_account_process(account,nil) unless account_not_synchronized_yet
 
       self.update_timestamps account
       
@@ -24,9 +24,10 @@ module ApiOperations
       def self.synchronize_account_process(account, user_map)
         # if there is a new user map
         if user_map
+          ApiOperations::Common.log(:debug,nil,"Started note synchronization for the new user map with id = " + user_map.id.to_s + " of the account with id = " + account.id.to_s)
+          
           begin
             # get the feed of changed notes per contact of this new user map from Ringio
-            ApiOperations::Common.log(:debug,nil,"Getting the changed notes for the new user map with id = " + user_map.id.to_s + " of the account with id = " + account.id.to_s)
             user_rg_feed = self.fetch_individual_user_rg_feed user_map
             # as it is the first synchronization for this user map, we are not interested in deleted notes
             rg_deleted_notes_ids = []
@@ -39,6 +40,8 @@ module ApiOperations
           rescue Exception => e
             ApiOperations::Common.log(:error,e,"\nProblem synchronizing the contacts of the new user map with id = " + um.id.to_s)
           end
+
+          ApiOperations::Common.log(:debug,nil,"Finished note synchronization for the new user map with id = " + user_map.id.to_s + " of the account with id = " + account.id.to_s)
         else
           begin
             # get the feed of changed notes per contact of this Ringio account from Ringio
@@ -175,27 +178,29 @@ module ApiOperations
 
 
       def self.synchronize_contact(individual,author_user_map, contact_map, contact_rg_feed, rg_deleted_notes_ids)
-        hr_updated_note_recordings = contact_map.hr_updated_note_recordings individual
-        # TODO: get true feeds of deleted notes (currently Highrise does not offer it)
-        hr_notes = contact_map.hr_notes
-
-        # get the deleted notes (those that don't appear anymore in the total)
-        hr_deleted_notes_ids = individual ? [] : contact_map.note_maps.reject{|nm| hr_notes.index{|hr_n| hr_n.id == nm.hr_note_id}}.map{|nm| nm.hr_note_id}
-
-        # give priority to Highrise: discard changes in Ringio to notes that have been changed in Highrise
-        self.purge_duplicated_changes(hr_updated_note_recordings,hr_deleted_notes_ids,contact_rg_feed,rg_deleted_notes_ids)
-
-        ApiOperations::Common.log(:debug,nil,"Started applying note changes from Ringio to Highrise for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
-        # apply changes from Ringio to Highrise
-        self.update_rg_to_hr(author_user_map,contact_map,contact_rg_feed)
-        self.delete_rg_to_hr(author_user_map,rg_deleted_notes_ids) unless individual
-        ApiOperations::Common.log(:debug,nil,"Finished applying note changes from Ringio to Highrise for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
-        
-        ApiOperations::Common.log(:debug,nil,"Started applying note changes from Highrise to Ringio for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
-        # apply changes from Highrise to Ringio
-        self.update_hr_to_rg(author_user_map,contact_map,hr_updated_note_recordings)
-        self.delete_hr_to_rg(author_user_map,hr_deleted_notes_ids) unless individual
-        ApiOperations::Common.log(:debug,nil,"Finished applying note changes from Highrise to Ringio for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
+        if contact_rg_feed.present? || rg_deleted_notes_ids.present?
+          hr_updated_note_recordings = contact_map.hr_updated_note_recordings individual
+          # TODO: get true feeds of deleted notes (currently Highrise does not offer it)
+          hr_notes = contact_map.hr_notes
+  
+          # get the deleted notes (those that don't appear anymore in the total)
+          hr_deleted_notes_ids = individual ? [] : contact_map.note_maps.reject{|nm| hr_notes.index{|hr_n| hr_n.id == nm.hr_note_id}}.map{|nm| nm.hr_note_id}
+  
+          # give priority to Highrise: discard changes in Ringio to notes that have been changed in Highrise
+          self.purge_duplicated_changes(hr_updated_note_recordings,hr_deleted_notes_ids,contact_rg_feed,rg_deleted_notes_ids)
+  
+          ApiOperations::Common.log(:debug,nil,"Started applying note changes from Ringio to Highrise for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
+          # apply changes from Ringio to Highrise
+          self.update_rg_to_hr(author_user_map,contact_map,contact_rg_feed)
+          self.delete_rg_to_hr(author_user_map,rg_deleted_notes_ids) unless individual
+          ApiOperations::Common.log(:debug,nil,"Finished applying note changes from Ringio to Highrise for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
+          
+          ApiOperations::Common.log(:debug,nil,"Started applying note changes from Highrise to Ringio for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
+          # apply changes from Highrise to Ringio
+          self.update_hr_to_rg(author_user_map,contact_map,hr_updated_note_recordings)
+          self.delete_hr_to_rg(author_user_map,hr_deleted_notes_ids) unless individual
+          ApiOperations::Common.log(:debug,nil,"Finished applying note changes from Highrise to Ringio for the contact map with id = " + contact_map.id.to_s + " by the author user map with id = " + author_user_map.id.to_s)
+        end
       end
